+++ src/com/android/bluetooth/opp/BluetoothOppObexServerSession.java
@@ -196,17 +196,17 @@ public class BluetoothOppObexServerSession extends ServerRequestHandler implemen
         private static final int sSleepTime = 500;
         private Uri contentUri;
         private Context mContext1;
-        private int position;
+        private long position;
         private volatile boolean interrupted = false;
 
-        public ContentResolverUpdateThread(Context context, Uri cntUri, int pos) {
+        public ContentResolverUpdateThread(Context context, Uri cntUri, long pos) {
             super("BtOpp Server ContentResolverUpdateThread");
             mContext1 = context;
             contentUri = cntUri;
             position = pos;
         }
 
-        public void updateProgress (int pos) {
+        public void updateProgress (long pos) {
             position = pos;
         }
 
@@ -379,7 +379,7 @@ public class BluetoothOppObexServerSession extends ServerRequestHandler implemen
         ContentValues values = new ContentValues();
 
         values.put(BluetoothShare.FILENAME_HINT, name);
-        values.put(BluetoothShare.TOTAL_BYTES, length.intValue());
+        values.put(BluetoothShare.TOTAL_BYTES, length);
         values.put(BluetoothShare.MIMETYPE, mimeType);
 
         if(mTransport instanceof BluetoothOppTransport){
@@ -585,7 +585,7 @@ public class BluetoothOppObexServerSession extends ServerRequestHandler implemen
             mContext.getContentResolver().update(contentUri, updateValues, null, null);
         }
 
-        int position = 0;
+        long position = 0;
         if (!error) {
             bos = new BufferedOutputStream(fileInfo.mOutputStream, 0x10000);
         }
